:art: moves Stripe/Tour/Callback folders into Misc
This commit moves Stripe/Tour/Callback folders into Misc to imporve structure
