+++ client/src/components/Containers/Profile/ProfileView.js
@@ -3,7 +3,7 @@ import { connect } from "react-redux";
 
 //Components
 // import UserModal from '../Modals/userModal';
-import StripeView from "../../Stripe/StripeView";
+import StripeView from "../../Misc/Stripe/StripeView";
 import { logout } from "../../../Auth/Auth";
 import Authentication from "../../Misc/authenticate/";
