+++ client/src/components/Misc/Stripe/StripeView.js
